Flow plot

Flow plot of connections

Usage

flows_sfn(result, lowcut = NULL, uppcut = NULL)

Arguments

result

SeaGraph object (i.e. output of seagraph) or sfnetwork or sf object containing 'from', 'to' and 'weight' column names.

lowcut

Optional percentile of threshold for not plotting connections below this value.

uppcut

Optional percentile of threshold for not plotting connections above this value.

Value

A leaflet flow map output.

Examples

graph_result <- seagraph(component_u    = get_component_u(),
                         component_v    = get_component_v(),
                         mask_shapefile = NULL,
                         k_neighbors    = 7)

flows_sfn(graph_result)

flows_sfn(graph_result, lowcut = 0.1, uppcut = 0.9)